% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txsos-package.R
\docType{package}
\name{txsos-package}
\alias{txsos}
\alias{txsos-package}
\title{txsos: Texas Secretary of State Precinct-Level Election Returns}
\description{
Written entirely by Cursor with guidance from Shiro Kuriwaki. Builds a precinct-level database of Texas election returns. Current functions cover: fetching precinct/VTD returns from the Capitol Data Portal ('fetch_ted_returns'); reshaping them to tidy long form ('tidy_ted_returns'); attaching CAGE candidate identifiers ('add_cage_keys'); building and storing a normalized precinct/candidate/votes schema as Parquet ('build_tables', 'write_returns_parquet', 'open_returns_parquet'); and summarizing and validating totals ('vote_share', 'candidate_totals', 'validate_totals').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kuriwaki/sos-state-tx-elections}
  \item Report bugs at \url{https://github.com/kuriwaki/sos-state-tx-elections/issues}
}

}
\author{
\strong{Maintainer}: Shiro Kuriwaki \email{shirokuriwaki@gmail.com}

Authors:
\itemize{
  \item Shiro Kuriwaki \email{shirokuriwaki@gmail.com}
}

}
\keyword{internal}
